Strips '{{~' and '~}}' from 2 of 3 md files
Removes '{{~' and '~}}' from pkg_binds.md and plan_helpers.md but leaves
them in upgrade/_index.md. The thought is that in the two files in which
they '{{~' and '~}}' were removed their presence were more likely to
cause issues for an end user than their absence if an the code was
copied and pasted without proper reflection upon the semantics of '{{~'
and '~}}'.  The usage in upgrade/_iindex.md is appropriate.

Signed-off-by: Jason Heath <jason.heath@progress.com>